// src/types/themeTypes/ThemeType.ts
export type ThemeTypes = {
  name: string;
  dark: boolean;
  variables?: object;
  colors: {
    primary?: string;
    primarydisabled?: string;
    secondary?: string;
    secondarydisabled?: string;
    info?: string;
    success?: string;
    accent?: string;
    warning?: string;
    error?: string;
    lightprimary?: string;
    lightprimarydisabled?: string;
    lightsecondary?: string;
    lightsecondarydisabled?: string;
    lightsuccess?: string;
    lighterror?: string;
    lightwarning?: string;
    darkprimary?: string;
    darksecondary?: string;
    darkText?: string;
    lightText?: string;
    borderLight?: string;
    inputBorder?: string;
    containerBg?: string;
    surface?: string;
    background?: string;
    'on-surface-variant'?: string;
    facebook?: string;
    twitter?: string;
    linkedin?: string;
    gray100?: string;
    primary200?: string;
    secondary200?: string;
    card?: string;
    main?: string;
    lightmain?: string;
    darkmain?: string;
    textmain?: string;
  };
};